Added a RCM-based structure n the server
