+++ server/server.js
@@ -1,12 +1,13 @@
 const app = require('express')();
 const server = require('http').createServer(app);
+const socketControllers = require('./controllers/socketController')
 const io = require('socket.io')(server, {
     cors: {
       origin: "*" // To remove Cross-origin error
     }
 });
 
-const router = require('./router');
+const router = require('./routes/router');
 
 const PORT = process.env.PORT || 5000
 
@@ -14,15 +15,7 @@ io.on('connection', (socket)=> {
     console.log('Socket is connected!!');
 
     socket.on("chat", (payload) => {
-        console.log(payload)
-        if(payload.room == ''){
-            const resObj = {message: "You dont have access to any room", userID: payload.userID, room: "", timeStamp: payload.timeStamp};
-            socket.emit("recieve-message", resObj );
-        }
-        else{
-            const resObj = {message: payload.message, userID: payload.userID, room: payload.userID, timeStamp: payload.timeStamp}
-            socket.broadcast.to(payload.room).emit("recieve-message", resObj);
-        }
+        socketControllers.handleChat(payload);
     })
 
     socket.on("join-room", room => {
@@ -30,10 +23,7 @@ io.on('connection', (socket)=> {
     })
 
     socket.on("get-online-users", room => {
-        console.log(room)
-        const onlineUsersArray = Array.from(io.sockets.adapter.rooms.get(room))
-        console.log(onlineUsersArray)
-        io.to(room).emit("recieve-online-users", onlineUsersArray)
+        socketControllers.getOnlineUsers(room);
     })
 
     socket.on('disconnect', () => {
